style: change navbar component, change image format from png to svg
