+++ package.json
@@ -32,7 +32,7 @@
     "@emotion/styled": "^11.11.0",
     "@google-cloud/storage": "^6.11.0",
     "@hookform/resolvers": "^3.1.1",
-    "@katitb2023/database": "git+https://github.com/KATITB2023/database.git#8bca40ca581a13eeed9f51555be4b4340c1349bd",
+    "@katitb2023/database": "git+https://github.com/KATITB2023/database.git#ebb596e1a226ec201ed2e0b6ca62e0b9be6c20ec",
     "@next-auth/prisma-adapter": "^1.0.5",
     "@opentelemetry/api": "^1.4.1",
     "@opentelemetry/exporter-trace-otlp-http": "^0.40.0",
